forecast_project/requirements.txt
flask
pandas
numpy
statsmodels
google-cloud-storage
gunicorn